Create interpret_date_assignment_for class method
Our usage of the interpret_date gem has created a pattern of defining
several assignment methods with the same logic inside of them which is
not helpful in the model and can clutter them with enough dates defined.
This commit introduces a class method that is included with
`InterpretDate` that allows you to define the attributes which need
their assignment interpreted by interpret date so we can clean up the
models.

[380933786845645](https://app.asana.com/0/380933786845645/380933786845645)