Add DateType to interpret date for rails dates
Our usage of the `interpret_date` gem has created a pattern of defining
several assignment methods with the same logic inside of them which is
not helpful in the model and can clutter them with enough dates defined.
This commit introduces a `DateType` class that can be used by the
ActiveRecord attributes API that automatically casts the type value
using `interpret_date` so we don't have to override Model attribute
writers in applications anymore.

[380933786845645](https://app.asana.com/0/380933786845645/380933786845645)